+++ src/Domain/Statistics/StaticWindowListener.java
@@ -4,7 +4,7 @@ import java.util.HashMap;
 
 public interface StaticWindowListener {
     void onTimeChange(double time);
-    void onHealthChange(int health);
+    void onHealthChange(double health);
     void onScoreChange(double score);
     void onAmmunitionChange(HashMap<String,HashMap<String,Integer>> map);
 }
